Splitting the code for cleaning-up
